Simplifying how direct sampling, filtering, offset tuning etc interact.
Treat direct sampling as a tuner with a LO of 0.

Disabled no-mod direct sampling; you can now get the same effect by
directly tuning now (between ~ 12-25Mhz)

Move responsibility for setting the IF filter entirely into the tuner;
it sets the filter as part of tuning, not as a separate operation, as
the target frequency is always known.