rtl_adsb: fix hanging upon a signal in Fedora 27
This code stayed unchanged for many years, but for some reason
rtl_adsb started hanging upon exit:

*b66116a5164b69281eacc42ae950;
^CSignal caught, exiting!
 <------ hangs here forever

Examining it with gdb reveals that the demod thread waits
peacefully on the condition variable, which we're trying to
destroy. Either the signals killed all threads before, or
condition variables were possible to destroy while other
threads still waited on them.

The easiest fix appears to be just cancel the demod thread
and wait for it to exit before proceeding for the door.